# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MC Fantasy Football Simulator API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/players": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "List Players",
        "operationId": "list_players_api_players_get",
        "parameters": [
          {
            "name": "position",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Position"
            }
          },
          {
            "name": "has_projection",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Has Projection"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerRow"
                  },
                  "title": "Response List Players Api Players Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/players/{player_id}/distribution": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "Get Distribution",
        "operationId": "get_distribution_api_players__player_id__distribution_get",
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          },
          {
            "name": "n",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20000,
              "minimum": 100,
              "default": 5000,
              "title": "N"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistributionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DistributionBody": {
        "properties": {
          "n_samples": {
            "type": "integer",
            "title": "N Samples"
          },
          "floor_p10": {
            "type": "number",
            "title": "Floor P10"
          },
          "median_p50": {
            "type": "number",
            "title": "Median P50"
          },
          "ceiling_p90": {
            "type": "number",
            "title": "Ceiling P90"
          },
          "mean": {
            "type": "number",
            "title": "Mean"
          },
          "std": {
            "type": "number",
            "title": "Std"
          },
          "histogram": {
            "$ref": "#/components/schemas/Histogram"
          }
        },
        "type": "object",
        "required": [
          "n_samples",
          "floor_p10",
          "median_p50",
          "ceiling_p90",
          "mean",
          "std",
          "histogram"
        ],
        "title": "DistributionBody"
      },
      "DistributionResponse": {
        "properties": {
          "player_id": {
            "type": "integer",
            "title": "Player Id"
          },
          "gsis_id": {
            "type": "string",
            "title": "Gsis Id"
          },
          "projection": {
            "type": "object",
            "title": "Projection"
          },
          "scoring_preset": {
            "type": "string",
            "enum": [
              "standard",
              "half_ppr",
              "full_ppr"
            ],
            "title": "Scoring Preset"
          },
          "computed_points": {
            "type": "number",
            "title": "Computed Points"
          },
          "distribution": {
            "$ref": "#/components/schemas/DistributionBody"
          },
          "fit": {
            "$ref": "#/components/schemas/FitInfo"
          }
        },
        "type": "object",
        "required": [
          "player_id",
          "gsis_id",
          "projection",
          "scoring_preset",
          "computed_points",
          "distribution",
          "fit"
        ],
        "title": "DistributionResponse"
      },
      "FitInfo": {
        "properties": {
          "fitted_at": {
            "type": "string",
            "title": "Fitted At"
          },
          "historical_seasons": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Historical Seasons"
          },
          "games_used": {
            "type": "integer",
            "title": "Games Used"
          }
        },
        "type": "object",
        "required": [
          "fitted_at",
          "historical_seasons",
          "games_used"
        ],
        "title": "FitInfo"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Histogram": {
        "properties": {
          "bin_edges": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Bin Edges"
          },
          "counts": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Counts"
          }
        },
        "type": "object",
        "required": [
          "bin_edges",
          "counts"
        ],
        "title": "Histogram"
      },
...
      "PlayerRow": {
        "properties": {
          "player_id": {
            "type": "integer",
            "title": "Player Id"
          },
          "gsis_id": {
            "type": "string",
            "title": "Gsis Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "team": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team"
          },
          "position": {
            "type": "string",
            "title": "Position"
          },
          "projected_stats": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Projected Stats"
          },
          "projected_points": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projected Points"
          },
          "adp_snake": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adp Snake"
          },
          "adp_auction": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adp Auction"
          }
        },
        "type": "object",
        "required": [
          "player_id",
          "gsis_id",
          "name",
          "team",
          "position"
        ],
        "title": "PlayerRow"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}